refactor(examples): split image viewer
