+++ tests/cycles/WF-165/tranche-c-families.ts
@@ -36,9 +36,12 @@ export const TRANCHE_C_FAMILY_MEMBERS = {
     'examples/image-viewer/image-viewer-options.ts',
     'examples/image-viewer/image-viewer-update.ts',
     'examples/image-viewer/image-viewer-files.ts',
+    'examples/image-viewer/image-viewer-picker.ts',
     'examples/image-viewer/image-viewer-render.ts',
     'examples/image-viewer/image-viewer-preview.ts',
+    'examples/image-viewer/image-viewer-load.ts',
     'examples/image-viewer/image-viewer-tuning.ts',
+    'examples/image-viewer/image-viewer-tuning-update.ts',
     'examples/image-viewer/image-viewer-create.ts',
     'examples/image-viewer/image-viewer-entry.ts',
   ],
